Moved utils package to util
Moved utils package to util in order to provide a clear separation for
later spliting the project into modules.
