+++ src/main/java/org/cirdles/topsoil/app/util/GetDocumentsDirectoryOperation.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.app.utils;
+package org.cirdles.topsoil.app.util;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.cirdles.utils.PlatformDependentOperation;
 
 /**
  * GetDocumentsDirectoryOperation.java
