+++ src/test/java/org/cirdles/topsoil/app/util/GetApplicationDirectoryOperationTest.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.cirdles.topsoil.app.utils;
+package org.cirdles.topsoil.app.util;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
